back/requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
django-cors-headers==4.1.0
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-flex-fields==1.0.2
drf-spectacular==0.26.3
drf-yasg==1.21.6
greenlet==2.0.2
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
kombu==5.3.1
packaging==23.1
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.38
psycopg2-binary==2.9.3
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
